monitor: use signed value for factor and offset
The code was using size_t for factor and offset. For some fans the
offset could be negative, so it needs to use signed types.
This commit changes the offset type to int64_t.

+++ monitor/tach_sensor.cpp
@@ -75,7 +75,7 @@ TachSensor::TachSensor(Mode mode,
                        size_t funcDelay,
                        const std::string& interface,
                        size_t factor,
-                       size_t offset,
+                       int64_t offset,
                        size_t timeout,
                        const sdeventplus::Event& event) :
     _bus(bus),
